producers/requirements.txt
websockets>=12.0
kafka-python-ng>=2.2.2
requests>=2.31.0